refactor(hooks): split read-sidecar from json-helper.cjs
Extract the read-sidecar operation into scripts/hooks/lib/sidecar-ops.cjs.
json-helper.cjs routes to the domain module before its main switch, keeping
the external CLI interface unchanged. This establishes the split pattern
for future extraction of learning-ops and json-io-ops.
